Add files via upload
These are the Jacobian matrix and the matrix Q for the inner product <a,b>_Q to obtain the adjoint matrix via inv(Q)*J'*Q. The Q matrix is basically the discrete cell volume in the diagonal. Leading dimension is 48715 of a mesh with 9743 control volumes. 
They are in Matlab .mat format using Matlab R2015a.